Move registerServiceWorker, notification to utils
- utils is a new directory under ./src/ holding misc modules that don't seem to
  fit anywhere else.
- createStore was excluded from utils since the store is a core part of our
  react/redux app, and so it does make sense as a standalone module in the root
  ./src/ directory
- Also fixed #37 unnecessary yield undefined.
